requirements.txt
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
urllib3<2
youtube-transcript-api==0.6.1
beautifulsoup4==4.12.3
streamlit==1.31.1
